pyproject.toml
[project]
name = "mini-node"
version = "0.9.0"
description = "GDI dataset services (FAIR Data Point v1.2, GA4GH Beacon v2)"
readme = "README.md"
license = "Unlicense"
license-files = [ "LICENSE" ]
authors = [
    { name = "Erik Jaaniso", email = "erik.jaaniso@ut.ee" },
    { name = "Martti Tamm", email = "martti.tamm@ut.ee" }
]

requires-python = ">=3.14,<4"

dependencies = [
    "fastapi [standard] (>0.124,<1.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "rdflib (>=7.5.0,<8.0.0)",
    "minio (>=7.2.20,<8.0.0)",
    "pyshacl (>=0.30.1,<0.31.0)",
    "python-json-logger (>=4.0.0,<5.0.0)",
    "watchdog (>=6.0.0,<7.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pyarrow (>=22.0.0,<23.0.0)",
    "isoduration (>=20.11.0,<21.0.0)",
    "cyvcf2 (>=0.31.4,<0.32.0)",
    "argparse (>=1.4.0,<2.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
black = "*"
poetry-core = "2.*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
